requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10.1
google-cloud-datastore==1.0.0
google-cloud-storage==1.0.0
six==1.10.0
base58==0.2.5
